fix jet charge variable type

+++ FlatTreeProducer/interface/FlatTree.hh
@@ -823,8 +823,8 @@ class FlatTree
    std::vector<float> jet_muonEnergy;
    std::vector<float> jet_photonEnergy;
 
-   std::vector<int> jet_charge;
-   std::vector<int> jet_chargeVec;
+   std::vector<float> jet_charge;
+   std::vector<float> jet_chargeVec;
    std::vector<int> jet_chargedMultiplicity;
    std::vector<int> jet_neutralMultiplicity;
    std::vector<int> jet_chargedHadronMultiplicity;
